Added socket configuration
A simple express server was added on port 3001. This server will handle
the socket connections with Socket.io lib.

Added the socket.io server on the express
and a socket.io on the client typeracer page. The following events were
added:

- if game has less than two players, it will add a new player then emit
a 'player.joined' event
- if on player join, the game have two players it will emit the
'game.can.start' event
- if the game has more than 2 players it will emit the 'game.is.full'
event